requirements.txt
streamlit==1.29.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
uuid==1.30
numpy==1.26.2
# Install CPU-only PyTorch using the official command:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
gunicorn==20.1.0